Transaction verification almost complete

+++ tests/transaction_tests.py
@@ -17,7 +17,8 @@ class TestTransactionClass(unittest.TestCase):
   
   def setUp(self):
     self.key = KeyStore.getPublicKey()
-    os.remove('db.db') # remove the database
+    if os.path.isfile('db.db'):
+      os.remove('db.db') # remove the database
     c = CoinBase() # give ourselves 100 to start with
     
   def test_create_transaction(self):
@@ -26,11 +27,12 @@ class TestTransactionClass(unittest.TestCase):
     trans.add_output(Transaction.Output(10, self.key)) # just pay to ourselves for now
     
     # verify the transaction
-    message = SHA256.new(str.encode('signature'))
-    verifier = PKCS1_v1_5.new(trans.output[0].pubKey)
+    #message = SHA256.new(str.encode('signature'))
+    #verifier = PKCS1_v1_5.new(trans.output[0].pubKey)
     
-    verified = verifier.verify(message, trans.input[0].signature)
-    self.assertTrue(verified)
+    #verified = verifier.verify(message, trans.input[0].signature)
+    
+    self.assertIsNotNone(trans)
     
   def test_build_raw_transaction(self):
 
@@ -51,6 +53,8 @@ class TestTransactionClass(unittest.TestCase):
     t.add_output(Transaction.Output(20, otherKey.publickey())) # give other 20
     t.finish_transaction()
     
+    t.verify()
+    
     # find all unspent outputs for ourselves and for other.
     db = DB()
     myKey = KeyStore.getPublicKey()
@@ -59,6 +63,34 @@ class TestTransactionClass(unittest.TestCase):
     
     self.assertEqual(myOuts[0].value, 80) # we should have 80 left
     self.assertEqual(othersOuts[0].value, 20) # other should have 20 now
+    
+  def test_verify_transactions(self):
+    otherKey = RSA.generate(2048)
+    t = Transaction()
+    t.add_output(Transaction.Output(20, otherKey.publickey()))
+    t.finish_transaction()
+    
+    verified = t.verify()
+    self.assertTrue(verified)
+    
+  def test_verify_transaction_fail(self):
+    otherKey = RSA.generate(2048)
+    myKey = RSA.generate(2048)
+    c = CoinBase(owner=myKey)
+    c.finish_transaction()
+    t = Transaction(owner=myKey)
+    t.add_output(Transaction.Output(20, otherKey.publickey()))
+    t.finish_transaction()
+    
+    # purposely modify the input and sign with other's key
+    i = t.input[0]
+    message = SHA256.new(b'MESSAGE')
+    signer = PKCS1_v1_5.new(otherKey)
+    signature = signer.sign(message)
+    i.signature = signature
+    
+    verified = t.verify()
+    self.assertFalse(verified)
     
     
 if __name__ == '__main__':
